// src/app/modules/course/course.interface.ts
import { Types } from "mongoose";

 
export type TPreRequisiteCourses = {
    course: Types.ObjectId;
    isDeleted: boolean;
    
}

export type TCourse = {
    title: string;
    prefix: string;
    code: number;
    credits: number;
    isDeleted?:boolean;
    preRequisiteCourses:[TPreRequisiteCourses] 
}


export type TCourseFaculty = {
    course: Types.ObjectId;
    faculties:[Types.ObjectId];
}